+++ src/main/java/io/github/astrapi69/crypt/data/key/writer/EncryptedPrivateKeyWriter.java
@@ -51,7 +51,7 @@ import javax.crypto.SecretKey;
 import javax.crypto.SecretKeyFactory;
 import javax.crypto.spec.PBEKeySpec;
 
-import io.github.astrapi69.crypt.api.algorithm.KeyPairGeneratorAlgorithm;
+import io.github.astrapi69.crypt.api.algorithm.key.KeyPairGeneratorAlgorithm;
 import io.github.astrapi69.crypt.api.compound.CompoundAlgorithm;
 import io.github.astrapi69.crypt.data.factory.AlgorithmParameterSpecFactory;
 import io.github.astrapi69.crypt.data.factory.SecretKeyFactoryExtensions;