+++ src/main/java/io/github/astrapi69/crypt/data/key/KeyStoreExtensions.java
@@ -33,7 +33,7 @@ import java.security.KeyStoreException;
 import java.security.NoSuchAlgorithmException;
 import java.security.cert.CertificateException;
 
-import io.github.astrapi69.crypt.api.algorithm.KeystoreType;
+import io.github.astrapi69.crypt.api.type.KeystoreType;
 import io.github.astrapi69.crypt.data.factory.KeyStoreFactory;
 
 /**